implement security for blog

+++ src/main/java/com/example/blog/service/RedisService.java
@@ -1,7 +1,7 @@
 package com.example.blog.service;
 
 public interface RedisService {
-    void setString(String key, String value,int expire);
+    void setString(String key, String value,long expire);
     String getString(String key);
     void deleteKey(String key);
     boolean setStringIfAbsent(String key, String value,int expire);
